Refactoring continues

+++ app/src/main/java/com/brightkey/nickfl/myutilities/entities/ConfigEntity.kt
@@ -8,9 +8,9 @@ class ConfigEntity {
     var vendorNameColor: String? = null   // #F58233
     var accountNumber: String? = null     // 1291-9-1-2938293
     var unitType: String? = null          // m3, kWh, ..
-    var unitPrice0: Double = 0.toDouble()        // $20.0
-    var unitPrice1: Double = 0.toDouble()        // $20.0
-    var unitPrice2: Double = 0.toDouble()        // $20.0
+    var unitPrice0: Double = 0.0          // $20.0
+    var unitPrice1: Double = 0.0          // $20.0
+    var unitPrice2: Double = 0.0          // $20.0
 
     override fun toString(): String {
         return "{utilityIcon:" + utilityIcon + ", utilityType:" + utilityType +
